requirements.txt
PyQt6>=6.5.0
pyzbar>=0.1.9
Pillow>=10.0.0

